+++ src/main/java/com/board/question/dto/QuestionsListDtoImpl.java
@@ -14,19 +14,19 @@ public class QuestionsListDtoImpl implements QuestionsListDto {
     private final LocalDateTime nowtime;
     private final String category;
     private final int view;
-    private final int replysCount;
+    private final Long replysCount;
     private final String nickname;
 
     public QuestionsListDtoImpl(Integer uploadnumber, String title, String content,
                                 LocalDateTime nowtime, String category, int view,
-                                long replysCount, String nickname) {
+                                Long replysCount, String nickname) {
         this.uploadnumber = uploadnumber;
         this.title = title;
         this.content = content;
         this.nowtime = nowtime;
         this.category = category;
         this.view = view;
-        this.replysCount = (int) replysCount;
+        this.replysCount = replysCount;
         this.nickname = nickname;
     }
 
@@ -36,6 +36,6 @@ public class QuestionsListDtoImpl implements QuestionsListDto {
     public LocalDateTime getNowtime() { return nowtime; }
     public String getCategory() { return category; }
     public int getView() { return view; }
-    public int getReplysCount() { return replysCount; }
+    public Long getReplysCount() { return replysCount; }
     public String getNickname() { return nickname; }
 }
